[bugfix] Redux 버그 수정중
- AddGroupLayout, GroupScreen 분리
